Various small changes to remove more warnings.
git-svn-id: http://svn.seqan.de/seqan/trunk@7882 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/mason/simulate_illumina.h
@@ -445,10 +445,10 @@ void buildSimulationInstructions(ReadSimulationInstruction<IlluminaReads> & inst
             if (inst.editString[i] == ERROR_TYPE_MISMATCH) {
                 // std::cout << "i == " << i << ", j == " << j << ", parameters.mismatchQualityMeans[j] == " << parameters.mismatchQualityMeans[j] << ", parameters.mismatchQualityStdDevs[j] == " << parameters.mismatchQualityStdDevs[j] << std::endl;
                 PDF<Normal> pdf(parameters.mismatchQualityMeans[j], parameters.mismatchQualityStdDevs[j]);
-                inst.qualities[i] = pickRandomNumber(rng, pdf);
+                inst.qualities[i] = static_cast<int>(pickRandomNumber(rng, pdf));
             } else {
                 PDF<Normal> pdf(parameters.qualityMeans[j], parameters.qualityStdDevs[j]);
-                inst.qualities[i] = pickRandomNumber(rng, pdf);
+                inst.qualities[i] = static_cast<int>(pickRandomNumber(rng, pdf));
             }
 
             if (inst.editString[i] == ERROR_TYPE_MISMATCH || inst.editString[i] == ERROR_TYPE_MATCH)
